feat: add requestAuthorizer
