+++ authorization-service/handler.js
@@ -1,44 +1,4 @@
-export const basicAuthorizer = async (event, ctx, cb) => {
-  console.log('basicAuthorizer: ', JSON.stringify(event));
+import { basicAuthorizer } from './handlers/basicAuthorizer';
+import { requestAuthorizer } from './handlers/requestAuthorizer';
 
-  if (event.type !== 'TOKEN') {
-    cb('Unauthorized');
-  }
-
-  try {
-    const authorizationToken = event.authorizationToken;
-    const encodedCreds = authorizationToken.split(' ')[1];
-    const buff = Buffer.from(encodedCreds, 'base64');
-    const plainCreds = buff.toString('utf-8').split(':');
-    const username = plainCreds[0];
-    const password = plainCreds[1];
-
-    console.log(`username: ${username} and password: ${password}`);
-
-    const storedUserPassword = process.env.uladzimiryeudakimovich;
-    const effect = !storedUserPassword || storedUserPassword !== password ? 'Deny' : 'Allow';
-
-    const policy = generatePolicy(encodedCreds, event.methodArn, effect);
-
-    cb(null, policy);
-  } catch (error) {
-    console.error('error: ', error);
-    cb(`Unauthorized: ${error.message}`);
-  }
-}
-
-const generatePolicy = (principalId, resource, effect = 'Deny') => {
-  return {
-    principalId: principalId,
-    policyDocument: {
-      Version: '2012-10-17',
-      Statement: [
-        {
-          Action: 'execute-api:Invoke',
-          Effect: effect,
-          Resource: resource
-        }
-      ]
-    }
-  };
-}
+export { basicAuthorizer, requestAuthorizer };
